refactor(camera): remove Camera component and change camera page to Client Page
