+++ src/app/diet/page.tsx
@@ -1,9 +1,61 @@
-import Camera from './_components/Camera';
+'use client';
+import { useEffect } from 'react';
+import { CameraView } from './_components/CameraView';
+import { PhotoView } from './_components/PhotoView';
+import { useCamera } from '@/hooks/useCamera';
+
+const containerStyle = 'x-[450px] h-[450px] md:w-[550px] md:h-[550px]';
 
 export default function CameraPage() {
+  const {
+    stream,
+    error,
+    photoData,
+    videoVisible,
+    videoRef,
+    canvasRef,
+    startCamera,
+    stopCamera,
+    takePhoto,
+    savePhoto,
+    setPhotoData,
+  } = useCamera(1.0);
+
+  useEffect(() => {
+    startCamera();
+    // cleanup function
+    return () => {
+      stopCamera();
+    };
+  }, []);
+
+  const handleTakePhoto = () => {
+    takePhoto();
+    if (photoData) {
+    }
+  };
+
+  const handleRetake = () => {
+    setPhotoData(null);
+    startCamera();
+  };
+
   return (
-    <main className='w-full flex flex-col justify-center items-center mt-8'>
-      <Camera />
+    <main className='bg-black w-full h-[calc(100vh-70px)] flex flex-col items-center'>
+      {error && <div className='text-red-500 p-2 bg-red-100 rounded'>에러: {error}</div>}
+
+      {stream && (
+        <div className={containerStyle}>
+          <CameraView videoRef={videoRef} isVideoVisible={videoVisible} onTakePhoto={handleTakePhoto} />
+        </div>
+      )}
+
+      <canvas ref={canvasRef} className='hidden' />
+      {photoData && (
+        <div className={containerStyle}>
+          <PhotoView photoData={photoData} onSave={savePhoto} onRetake={handleRetake} />
+        </div>
+      )}
     </main>
   );
 }
